refactor(e2e): deduplicate setupPlanDetailsPage and tighten URL pattern
- Extract setupPlanDetailsPage into a shared helper
  (testing/playwright/fixtures/helpers/planDetailsHelpers.ts) so
  plan-details-page.spec.ts and plan-details-resources.spec.ts share
  a single source of truth for plan navigation setup
- Tighten /\/resources/ URL regex to /\/resources(?:\?|$)/ so it
  only matches the segment at the end of the path

Resolves: MTV-5758
